# server/example-client/java/generated/code/api/openapi.yaml
openapi: 3.0.0
info:
  title: vcp_server
  version: 0.1.0
servers:
- url: /
paths:
  /vcp/createSignerData:
    post:
      description: Returns SignerData.
      operationId: createSignerData
      parameters:
      - description: "Which ZKP library to use. AC2C_BBS, AC2C_PS or DNC (error if\
          \ missing)."
        explode: true
        in: query
        name: zkpLib
        required: false
        schema:
          description: "Which ZKP library to use. AC2C_BBS, AC2C_PS or DNC (error\
            \ if missing)."
          nullable: true
          type: string
        style: form
      - description: A seed for a random number generator. Defaults to zero if not
          given.
        explode: true
        in: query
        name: rngSeed
        required: false
        schema:
          description: A seed for a random number generator. Defaults to zero if not
            given.
          format: uint64
          minimum: 0.0
          nullable: true
          type: integer
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSignerDataRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignerData'
          description: ""
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ""
      summary: Create the secret and public data used to sign and verify credentials.
      x-content-type: application/json
      x-accepts:
      - application/json
  /vcp/createBlindSigningInfo:
    post:
      description: Returns BlindSigningInfo
      operationId: createBlindSigningInfo
      parameters:
      - description: "Which ZKP library to use. AC2C_BBS, AC2C_PS or DNC (error if\
          \ missing)."
        explode: true
        in: query
        name: zkpLib
        required: false
        schema:
          description: "Which ZKP library to use. AC2C_BBS, AC2C_PS or DNC (error\
            \ if missing)."
          nullable: true
          type: string
        style: form
      - description: A seed for a random number generator. Defaults to zero if not
          given.
        explode: true
        in: query
        name: rngSeed
        required: false
        schema:
          description: A seed for a random number generator. Defaults to zero if not
            given.
          format: uint64
          minimum: 0.0
          nullable: true
          type: integer
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBlindSigningInfoRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlindSigningInfo'
          description: ""
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ""
      summary: Create BlindSigningInfo
      x-content-type: application/json
      x-accepts:
      - application/json
  /vcp/createAccumulatorData:
    post:
      description: Returns CreateAccumulatorResponse.
      operationId: createAccumulatorData
      parameters:
      - description: "Which ZKP library to use. AC2C_BBS, AC2C_PS or DNC (error if\
          \ missing)."
        explode: true
        in: query
        name: zkpLib
        required: false
        schema:
          description: "Which ZKP library to use. AC2C_BBS, AC2C_PS or DNC (error\
            \ if missing)."
          nullable: true
          type: string
        style: form
      - description: A seed for a random number generator. Defaults to zero if not
          given.
        explode: true
        in: query
        name: rngSeed
        required: false
        schema:
          description: A seed for a random number generator. Defaults to zero if not
            given.
          format: uint64
          minimum: 0.0
          nullable: true
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAccumulatorResponse'
          description: ""
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ""
      summary: Create an accumulator and its associated secret/public data.
      x-accepts:
      - application/json
  /vcp/sign:
    post:
      description: Returns Signature.
      operationId: sign
      parameters:
      - description: "Which ZKP library to use. AC2C_BBS, AC2C_PS or DNC (error if\
          \ missing)."
        explode: true
        in: query
        name: zkpLib
        required: false
        schema:
          description: "Which ZKP library to use. AC2C_BBS, AC2C_PS or DNC (error\
            \ if missing)."
          nullable: true
          type: string
        style: form
      - description: A seed for a random number generator. Defaults to zero if not
          given.
        explode: true
        in: query
        name: rngSeed
        required: false
        schema:
          description: A seed for a random number generator. Defaults to zero if not
            given.
          format: uint64
          minimum: 0.0
          nullable: true
          type: integer
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Signature'
          description: ""
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ""
      summary: Create a signature from the given values and SignerData.
      x-content-type: application/json
      x-accepts:
      - application/json
  /vcp/signWithBlindedAttributes:
    post:
      description: Returns a BlindSignature.
      operationId: signWithBlindedAttributes
      parameters:
      - description: "Which ZKP library to use. AC2C_BBS, AC2C_PS or DNC (error if\
          \ missing)."
        explode: true
        in: query
        name: zkpLib
        required: false
        schema:
          description: "Which ZKP library to use. AC2C_BBS, AC2C_PS or DNC (error\
            \ if missing)."
          nullable: true
          type: string
        style: form
      - description: A seed for a random number generator. Defaults to zero if not
          given.
        explode: true
        in: query
        name: rngSeed
        required: false
        schema:
          description: A seed for a random number generator. Defaults to zero if not
            given.
          format: uint64
          minimum: 0.0
          nullable: true
          type: integer
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignWithBlindedAttributesRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlindSignature'
          description: ""
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ""
      summary: "Create a BlindSignature from the given non-blinded values, blinding\
        \ info and SignerData."
      x-content-type: application/json
      x-accepts:
      - application/json
  /vcp/unblindBlindedSignature:
    post:
      description: Returns Signature.
      operationId: unblindBlindedSignature
      parameters:
      - description: "Which ZKP library to use. AC2C_BBS, AC2C_PS or DNC (error if\
          \ missing)."
        explode: true
        in: query
        name: zkpLib
        required: false
        schema:
          description: "Which ZKP library to use. AC2C_BBS, AC2C_PS or DNC (error\
            \ if missing)."
          nullable: true
          type: string
        style: form
      - description: A seed for a random number generator. Defaults to zero if not
          given.
        explode: true
        in: query
        name: rngSeed
        required: false
        schema:
          description: A seed for a random number generator. Defaults to zero if not
            given.
          format: uint64
          minimum: 0.0
          nullable: true
          type: integer
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnblindBlindedSignatureRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Signature'
          description: ""
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ""
      summary: Unblinded a blinded signature.
      x-content-type: application/json
      x-accepts:
      - application/json
  /vcp/createAccumulatorElement:
    post:
      description: Returns AccumulatorElement.
      operationId: createAccumulatorElement
      parameters:
      - description: "Which ZKP library to use. AC2C_BBS, AC2C_PS or DNC (error if\
          \ missing)."
        explode: true
        in: query
        name: zkpLib
        required: true
        schema:
          description: "Which ZKP library to use. AC2C_BBS, AC2C_PS or DNC (error\
            \ if missing)."
          type: string
        style: form
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccumulatorElement'
          description: ""
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ""
      summary: Create an accumulator element from the given text.
      x-content-type: application/json
      x-accepts:
      - application/json
...
components:
  schemas:
    SignerData:
      description: Contains a Signer's secret and public data.
      example:
        signerSecretData: signerSecretData
        signerPublicData:
          signerPublicSetupData: signerPublicSetupData
          signerPublicSchema:
          - CTText
          - CTText
          signerBlindedAttrIdxs:
          - 0
          - 0
      properties:
        signerPublicData:
          $ref: '#/components/schemas/SignerPublicData'
        signerSecretData:
          description: A Signer's secret keys.
          type: string
      required:
      - signerPublicData
      - signerSecretData
      type: object
...
    Error:
      example:
        reason: reason
        location: location
      properties:
        reason:
          type: string
        location:
          type: string
      required:
      - location
      - reason
      type: object
    CreateSignerDataRequest:
      description: Supplies the schema and the (possibly empty) indices of attributes
        to be blinded.
      example:
        claimTypes:
        - CTText
        - CTText
        blindedAttributeIndices:
        - 0
        - 0
      properties:
        claimTypes:
          description: The schema.
          items:
            $ref: '#/components/schemas/ClaimType'
          type: array
        blindedAttributeIndices:
          description: The attributes to be blinded.
          items:
            format: uint64
            minimum: 0.0
            type: integer
          type: array
      required:
      - blindedAttributeIndices
      - claimTypes
      type: object
...
    CreateAccumulatorResponse:
      description: Contains the AccumulatorData and the Accumulator.
      example:
        accumulatorData:
          accumulatorPublicData: accumulatorPublicData
          accumulatorSecretData: accumulatorSecretData
        accumulator: accumulator
      properties:
        accumulatorData:
          $ref: '#/components/schemas/AccumulatorData'
        accumulator:
          description: The value of an accumulator.
          type: string
      required:
      - accumulator
      - accumulatorData
      type: object
...